+++ client/CHANGELOG.md
@@ -1,3 +1,7 @@
+# Version 0.0.6
+- Migrate to GitHub repository.
+- Publish to VSCode Marketplace.
+
 # Version 0.0.4
 - Added "Show operation duration" command
 - Added virtual document with gid report. Use "Show log entries for the gid" command to generate it.